requirements.dev.txt
anyascii==0.3.2 ; python_version >= "3.12" and python_version < "4.0"
argcomplete==3.3.0 ; python_version >= "3.12" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4.0"
beautifulsoup4==4.12.3 ; python_version >= "3.12" and python_version < "4.0"
brotli==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.12" and python_version < "4.0"
cfgv==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0"
commitizen==3.27.0 ; python_version >= "3.12" and python_version < "4.0"
decli==0.6.2 ; python_version >= "3.12" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.12" and python_version < "4.0"
distlib==0.3.8 ; python_version >= "3.12" and python_version < "4.0"
django-debug-toolbar==4.3.0 ; python_version >= "3.12" and python_version < "4.0"
django-filter==24.2 ; python_version >= "3.12" and python_version < "4.0"
django-modelcluster==6.3 ; python_version >= "3.12" and python_version < "4.0"
django-permissionedforms==0.1 ; python_version >= "3.12" and python_version < "4.0"
django-taggit==5.0.1 ; python_version >= "3.12" and python_version < "4.0"
django-treebeard==4.7.1 ; python_version >= "3.12" and python_version < "4.0"
django==5.0.6 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework==3.15.1 ; python_version >= "3.12" and python_version < "4.0"
djhtml==3.0.6 ; python_version >= "3.12" and python_version < "4.0"
draftjs-exporter==5.0.0 ; python_version >= "3.12" and python_version < "4.0"
environs==11.0.0 ; python_version >= "3.12" and python_version < "4.0"
et-xmlfile==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
filelock==3.14.0 ; python_version >= "3.12" and python_version < "4.0"
filetype==1.2.0 ; python_version >= "3.12" and python_version < "4.0"
gunicorn==22.0.0 ; python_version >= "3.12" and python_version < "4.0"
hiredis==2.3.2 ; python_version >= "3.12" and python_version < "4.0"
identify==2.5.36 ; python_version >= "3.12" and python_version < "4.0"
idna==3.7 ; python_version >= "3.12" and python_version < "4.0"
importlib-metadata==7.1.0 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
l18n==2021.3 ; python_version >= "3.12" and python_version < "4.0"
laces==0.1.1 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
marshmallow==3.21.2 ; python_version >= "3.12" and python_version < "4.0"
nodeenv==1.8.0 ; python_version >= "3.12" and python_version < "4.0"
openpyxl==3.1.2 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.12" and python_version < "4.0"
pillow-heif==0.16.0 ; python_version >= "3.12" and python_version < "4.0"
pillow==10.3.0 ; python_version >= "3.12" and python_version < "4.0"
platformdirs==4.2.2 ; python_version >= "3.12" and python_version < "4.0"
pre-commit==3.7.1 ; python_version >= "3.12" and python_version < "4.0"
prompt-toolkit==3.0.36 ; python_version >= "3.12" and python_version < "4.0"
psycopg-binary==3.1.19 ; implementation_name != "pypy" and python_version >= "3.12" and python_version < "4.0"
psycopg[binary]==3.1.19 ; python_version >= "3.12" and python_version < "4.0"
pymemcache==4.0.0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
questionary==2.0.1 ; python_version >= "3.12" and python_version < "4.0"
redis[hiredis]==5.0.4 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.2 ; python_version >= "3.12" and python_version < "4.0"
ruff==0.4.5 ; python_version >= "3.12" and python_version < "4.0"
setuptools==70.0.0 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
soupsieve==2.5 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.0 ; python_version >= "3.12" and python_version < "4.0"
telepath==0.3.1 ; python_version >= "3.12" and python_version < "4.0"
termcolor==2.4.0 ; python_version >= "3.12" and python_version < "4.0"
tomlkit==0.12.5 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.11.0 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.2.1 ; python_version >= "3.12" and python_version < "4.0"
virtualenv==20.26.2 ; python_version >= "3.12" and python_version < "4.0"
wagtail==6.1.1 ; python_version >= "3.12" and python_version < "4.0"
wcwidth==0.2.13 ; python_version >= "3.12" and python_version < "4.0"
whitenoise[brotli]==6.6.0 ; python_version >= "3.12" and python_version < "4.0"
willow[heif]==1.8.0 ; python_version >= "3.12" and python_version < "4.0"
zipp==3.18.2 ; python_version >= "3.12" and python_version < "4.0"
